Cap which bin we can allocate in to the largest available
This method decides which bin we're actually going to allocate objects
in depending on whether they're powers of two or not.

Our bins are divided up by powers of 2, with the lower end being
inclusive. We only allocate in a particular bin if the requested size is
actually on the lower bound so that we can guarantee that it'll fit in
the first available free region, otherewise we allocate in the next
highest bin. This is an optimisation to avoid having to scan the regions
within each bin to find one that is the right size.

This commit prevents us from attempting to allocate regions larger than
256 slots in a bin that doesn't exist (Max bin + 1)