scan the regions to find one that fits
Previously we avoided this because we'd always allocate in one region
larger than the one that we required. But we had a bug where in the
largest case, we'd attempt to allocate into a bin larger than the
largest available, which would crash, so instead we're allocating into
the largest bin now, which means we still need to do this scan in some
cases.

On the plus side, for all allocation in bins that are not the largest,
this will always match on the first iteration