Make the max payload len 256
We can't guarantee sizing be available for payload lengths greater than
this so for now let's cap at 256

An alternative approach would be to special case 256 > n < 409 and
request a new page each time. But for now this is the simplest thing to
do until we can get some benchmarks and some numbers for how often this
is happening.